feat: add proptest property-based tests across forge-types and forge-worldgen
- forge-types/action.rs: 4 proptest! blocks covering discrete roundtrip,
  Communicate roundtrip, out-of-range returns None, and space_size floor
- forge-types/config.rs: 2 proptest! blocks for validate_config invariant
  and TOML serialization roundtrip across arbitrary valid dimensions/seeds
- forge-worldgen/lib.rs: 2 proptest! blocks for generation determinism and
  position in-bounds invariant across arbitrary seeds and small dimensions

+++ crates/forge-types/src/config.rs
@@ -820,11 +820,14 @@ num_agents = 4
 
             /// A `ForgeConfig` serialized to TOML and deserialized must produce
             /// identical `world.width`, `world.height`, and `world.seed` values.
+            ///
+            /// Seeds are restricted to `[0, i64::MAX]` because the `toml` v0.8 crate
+            /// serializes `u64` values via `i64` and rejects values that would overflow.
             #[test]
             fn prop_config_toml_roundtrip(
                 width  in valid_dimension(),
                 height in valid_dimension(),
-                seed   in 0u64..=u64::MAX,
+                seed   in 0u64..=(i64::MAX as u64),
             ) {
                 let mut config = ForgeConfig::default();
                 config.world.width  = width;
